updated the test drivers for octagons for long input data

+++ octagons/tests/libFuzzer/test_oct.c
@@ -6,8 +6,8 @@
 #include <stdio.h>
 #include <math.h>
 
-ap_linexpr0_t * create_linexpr0(int dim, int v1, int v2, int coeff1,
-		int coeff2, int scalar_value) {
+ap_linexpr0_t * create_linexpr0(long dim, long v1, long v2, long coeff1,
+		long coeff2, long scalar_value) {
 	ap_coeff_t *cst, *coeff;
 	ap_linexpr0_t * linexpr0 = ap_linexpr0_alloc(AP_LINEXPR_SPARSE, 2);
 	cst = &linexpr0->cst;
@@ -29,10 +29,10 @@ ap_linexpr0_t * create_linexpr0(int dim, int v1, int v2, int coeff1,
 	return linexpr0;
 }
 
-bool create_constraints(ap_lincons0_array_t *lincons0, int dim,
-		const int *data, size_t dataSize, unsigned int *dataIndex, FILE *fp) {
+bool create_constraints(ap_lincons0_array_t *lincons0, long dim,
+		const long *data, size_t dataSize, unsigned int *dataIndex, FILE *fp) {
 	size_t i;
-	int nbcons = MIN_NBCONS;
+	long nbcons = MIN_NBCONS;
 	if (!make_fuzzable(&nbcons, sizeof(nbcons), data, dataSize, dataIndex)) //number of constraints
 			{
 		return false;
@@ -41,7 +41,7 @@ bool create_constraints(ap_lincons0_array_t *lincons0, int dim,
 			nbcons >= MIN_NBCONS && nbcons <= MAX_NBCONS && nbcons >= dim)) {
 		return false;
 	}
-	fprintf(fp, "Number of constraints: %d\n", nbcons);
+	fprintf(fp, "Number of constraints: %ld\n", nbcons);
 	fflush(fp);
 	*lincons0 = ap_lincons0_array_make(nbcons);
 	for (i = 0; i < nbcons; i++) {
@@ -58,8 +58,8 @@ bool create_constraints(ap_lincons0_array_t *lincons0, int dim,
 		fflush(fp);
 		lincons0->p[i].constyp = type;
 
-		int fuzzableValues[5];
-		if (!make_fuzzable(fuzzableValues, 5 * sizeof(int), data, dataSize,
+		long fuzzableValues[5];
+		if (!make_fuzzable(fuzzableValues, 5 * sizeof(long), data, dataSize,
 				dataIndex)) {
 			return false;
 		}
@@ -82,7 +82,7 @@ bool create_constraints(ap_lincons0_array_t *lincons0, int dim,
 		if (!assume_fuzzable(fuzzableValues[4] > 0)) {
 			return false;
 		}
-		fprintf(fp, "Values: %d, %d, %d, %d, %d\n", fuzzableValues[0],
+		fprintf(fp, "Values: %ld, %ld, %ld, %ld, %ld\n", fuzzableValues[0],
 				fuzzableValues[1], fuzzableValues[2], fuzzableValues[3],
 				fuzzableValues[4]);
 		fflush(fp);
@@ -95,7 +95,7 @@ bool create_constraints(ap_lincons0_array_t *lincons0, int dim,
 }
 
 bool create_octagon(oct_t** octagon, ap_manager_t* man, oct_t * top,
-		int dim, const int *data, size_t dataSize, unsigned int *dataIndex,
+		long dim, const long *data, size_t dataSize, unsigned int *dataIndex,
 		FILE *fp) {
 	ap_lincons0_array_t constraints;
 	if (!create_constraints(&constraints, dim, data, dataSize, dataIndex, fp)) {
@@ -105,9 +105,9 @@ bool create_octagon(oct_t** octagon, ap_manager_t* man, oct_t * top,
 	return true;
 }
 
-bool make_fuzzable(void *array, size_t size, const int *data, size_t dataSize,
+bool make_fuzzable(void *array, size_t size, const long *data, size_t dataSize,
 		unsigned int *dataIndex) {
-	int numberOfElements = dataSize / sizeof(int);
+	int numberOfElements = dataSize / sizeof(long);
 	if (numberOfElements <= *dataIndex) {
 		return false;
 	}
@@ -124,11 +124,11 @@ bool assume_fuzzable(bool condition) {
 	return condition;
 }
 
-bool make_fuzzable_dimension(int *dim, const int *data, size_t dataSize,
+bool make_fuzzable_dimension(long *dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp) {
-	if (make_fuzzable(dim, sizeof(int), data, dataSize, dataIndex)) {
+	if (make_fuzzable(dim, sizeof(long), data, dataSize, dataIndex)) {
 		if (assume_fuzzable(*dim > MIN_DIM && *dim < MAX_DIM)) {
-			fprintf(fp, "Dim: %d\n", *dim);
+			fprintf(fp, "Dim: %ld\n", *dim);
 			fflush(fp);
 			return true;
 		}
